maint: relocating pure function to config
Calling config.IsLegacyAPIKey() does not require anything from a real
implementation of the Config interface. Move the pure function to the
file that defines the interface.
