+++ config/config_test.go
@@ -8,6 +8,7 @@ import (
 	"testing"
 	"time"
 
+	"github.com/honeycombio/husky/otlp"
 	"github.com/honeycombio/refinery/config"
 	"github.com/honeycombio/refinery/internal/configwatcher"
 	"github.com/honeycombio/refinery/logger"
@@ -1201,4 +1202,81 @@ func TestHealthCheckTimeout(t *testing.T) {
 		})
 	}
 
+}
+
+func TestIsLegacyKey(t *testing.T) {
+	testCases := []struct {
+		name     string
+		key      string
+		expected bool
+	}{
+		// 32-character classic API keys (hex digits only)
+		{name: "valid 32-char classic key - all lowercase", key: "a1b2c3d4e5f67890abcdef1234567890", expected: true},
+		{name: "valid 32-char classic key - all numbers", key: "12345678901234567890123456789012", expected: true},
+		{name: "valid 32-char classic key - all lowercase letters", key: "abcdefabcdefabcdefabcdefabcdefab", expected: true},
+		{name: "valid 32-char classic key - mixed hex", key: "0123456789abcdef0123456789abcdef", expected: true},
+		{name: "invalid 32-char key - uppercase letters", key: "A1B2C3D4E5F67890ABCDEF1234567890", expected: false},
+		{name: "invalid 32-char key - contains g", key: "a1b2c3d4e5f67890abcdefg234567890", expected: false},
+		{name: "invalid 32-char key - contains special chars", key: "a1b2c3d4e5f67890abcdef123456789!", expected: false},
+		{name: "invalid 32-char key - contains space", key: "a1b2c3d4e5f67890abcdef12345 7890", expected: false},
+
+		// 64-character classic ingest keys (pattern: ^hc[a-z]ic_[0-9a-z]*$)
+		{name: "valid 64-char ingest key - hcaic", key: "hcaic_1234567890123456789012345678901234567890123456789012345678", expected: true},
+		{name: "valid 64-char ingest key - hcbic", key: "hcbic_abcdef1234567890abcdef1234567890abcdef1234567890abcdef1234", expected: true},
+		{name: "valid 64-char ingest key - hczic", key: "hczic_0123456789abcdef0123456789abcdef0123456789abcdef0123456789", expected: true},
+		{name: "valid 64-char ingest key - mixed", key: "hcxic_1234567890123456789012345678901234567890123456789012345678", expected: true},
+		{name: "invalid 64-char ingest key - wrong prefix", key: "hc1ic_1234567890123456789012345678901234567890123456789012345678", expected: false},
+		{name: "invalid 64-char ingest key - uppercase in prefix", key: "hcAic_1234567890123456789012345678901234567890123456789012345678", expected: false},
+		{name: "invalid 64-char ingest key - missing underscore", key: "hcaic1234567890123456789012345678901234567890123456789012345678", expected: false},
+		{name: "invalid 64-char ingest key - uppercase in suffix", key: "hcaic_1234567890123456789012345678901234567890123456789012345A78", expected: false},
+		{name: "invalid 64-char ingest key - special char in suffix", key: "hcaic_123456789012345678901234567890123456789012345678901234567!", expected: false},
+
+		// Edge cases for length
+		{name: "empty key", key: "", expected: false},
+		{name: "too short - 31 chars", key: "a1b2c3d4e5f67890abcdef123456789", expected: false},
+		{name: "too long - 33 chars", key: "a1b2c3d4e5f67890abcdef12345678901", expected: false},
+		{name: "too short - 63 chars", key: "hcaic_123456789012345678901234567890123456789012345678901234567", expected: false},
+		{name: "too long - 65 chars", key: "hcaic_12345678901234567890123456789012345678901234567890123456789", expected: false},
+
+		// Non-classic keys (E&S keys should return false)
+		{name: "E&S key", key: "abc123DEF456ghi789jklm", expected: false},
+		{name: "E&S ingest key", key: "hcxik_1234567890123456789012345678901234567890123456789012345678", expected: false},
+
+		// Invalid patterns
+		{name: "random string", key: "this-is-not-a-key", expected: false},
+		{name: "numbers only but wrong length", key: "123456789012", expected: false},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			result := config.IsLegacyAPIKey(tc.key)
+			assert.Equal(t, tc.expected, result, "Expected IsLegacyApiKey(%q) to return %v", tc.key, tc.expected)
+		})
+	}
+}
+
+func BenchmarkIsLegacyAPIKey(b *testing.B) {
+	tests := []struct {
+		name string
+		key  string
+	}{
+		{"Valid classic key", "a1b2c3d4e5f67890abcdef1234567890"},
+		{"Invalid classic key", "abcdef0123456789abcdef01234567zz"},
+		{"Valid ingest key", "hcaic_1234567890123456789012345678901234567890123456789012345678"},
+		{"Invalid ingest key", "hcaic_1234567890123456789012345678901234567890123456789012345678"},
+	}
+
+	for _, tt := range tests {
+		b.Run(tt.name, func(b *testing.B) {
+			for i := 0; i < b.N; i++ {
+				_ = config.IsLegacyAPIKey(tt.key)
+			}
+		})
+
+		b.Run(tt.name+"/husky", func(b *testing.B) {
+			for i := 0; i < b.N; i++ {
+				_ = otlp.IsClassicApiKey(tt.key)
+			}
+		})
+	}
 }
